Implementing type conversions for the RuntimeContainer
The mixin stages in the RuntimeContainer are now labeled from 0 to
n-1, this level identifier type is used to created a reduced mixin
type with a fold operation. The reduced type can be used to cast
to different levels of the mixin.

Note: interface function are no longer virtual in order to allow
the cast to different mixin levels.